+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Error message for tests that should panic but pass
 
+### Cast
+
+#### Added
+
+- `-p|--package` flag to specify scarb package to work with
+
 ## [0.14.0] - 2024-01-11
 
 ### Forge